+++ scripts/__dropins__/storefront-requisition-list/chunks/updateRequisitionList.js
@@ -1,14 +1,6 @@
 /*! Copyright 2026 Adobe
 All Rights Reserved. */
-import{f as p,h as d}from"./fetch-graphql.js";import{events as I}from"@dropins/tools/event-bus.js";const m=`
-fragment REQUISITION_LIST_FRAGMENT on RequisitionList {
-    uid
-    name
-    description
-    items_count
-    updated_at
-  }
-`,f=`
+import{R as c,t as I}from"./transform-requisition-list.js";import{h as m}from"./fetch-error.js";import{events as d}from"@dropins/tools/event-bus.js";import{f as T}from"./fetch-graphql.js";const R=`
 fragment REQUISITION_LIST_ITEMS_FRAGMENT on RequistionListItems {
   items {
     uid
@@ -101,7 +93,7 @@ fragment REQUISITION_LIST_ITEMS_FRAGMENT on RequistionListItems {
     total_pages
   }
 }
-`;function b(t){var i,u;return t?{uid:t.uid,name:t.name,description:t.description,updated_at:t.updated_at,items_count:t.items_count,items:g((i=t.items)==null?void 0:i.items),page_info:(u=t.items)==null?void 0:u.page_info}:null}function g(t){return t!=null&&t.length?t.map(i=>{var _,a,s,r;const u={uid:i.uid,sku:(_=i.product)==null?void 0:_.sku,quantity:i.quantity,stock_status:((a=i.product)==null?void 0:a.stock_status)||"IN_STOCK",only_x_left_in_stock:((s=i.product)==null?void 0:s.only_x_left_in_stock)??null,customizable_options:i.customizable_options?i.customizable_options.map(e=>({uid:e.customizable_option_uid,is_required:e.is_required,label:e.label,sort_order:e.sort_order,type:e.type,values:e.values.map(n=>({uid:n.customizable_option_value_uid,label:n.label,price:n.price,value:n.value}))})):[],bundle_options:i.bundle_options||[],configurable_options:i.configurable_options?i.configurable_options.map(e=>({option_uid:e.configurable_product_option_uid,option_label:e.option_label,value_uid:e.configurable_product_option_value_uid,value_label:e.value_label})):[],samples:i.samples?i.samples.map(e=>({url:e.sample_url,sort_order:e.sort_order,title:e.title})):[],gift_card_options:i.gift_card_options||{}};return(r=i.configured_product)!=null&&r.name?{...u,configured_product:i.configured_product}:u}):[]}const T=`
+`,L=`
   mutation UPDATE_REQUISITION_LIST_MUTATION(
       $requisitionListUid: ID!,
       $name: String!,
@@ -124,7 +116,7 @@ fragment REQUISITION_LIST_ITEMS_FRAGMENT on RequistionListItems {
       }
     }
   }
-${m}
-${f}
-`,S=async(t,i,u,_,a,s)=>{var l,c;const{errors:r,data:e}=await p(T,{variables:{requisitionListUid:t,name:i,description:u,pageSize:_,currentPage:a}});if(r)return d(r);if(!((l=e==null?void 0:e.updateRequisitionList)!=null&&l.requisition_list))return null;const n=e.updateRequisitionList.requisition_list;let o=b(n);return(c=o==null?void 0:o.items)!=null&&c.length&&s&&(o={...o,items:await s(o.items)}),I.emit("requisitionList/data",o),o};export{m as R,f as a,b as t,S as u};
+${c}
+${R}
+`,E=async(r,a,u,_,l,t)=>{var n,o;const{errors:s,data:e}=await T(L,{variables:{requisitionListUid:r,name:a,description:u,pageSize:_,currentPage:l}});if(s)return m(s);if(!((n=e==null?void 0:e.updateRequisitionList)!=null&&n.requisition_list))return null;const p=e.updateRequisitionList.requisition_list;let i=I(p);return(o=i==null?void 0:i.items)!=null&&o.length&&t&&(i={...i,items:await t(i.items)}),d.emit("requisitionList/data",i),i};export{R,E as u};
 //# sourceMappingURL=updateRequisitionList.js.map