+++ scripts/__dropins__/storefront-requisition-list/components/index.d.ts
@@ -26,4 +26,5 @@ export * from './BatchActions/BatchActions';
 export * from './PageSizePicker/PageSizePicker';
 export * from './PaginationItemsCounter/PaginationItemsCounter';
 export * from './RequisitionListPicker/RequisitionListPicker';
+export * from './ShareRequisitionListContent/ShareRequisitionListContent';
 //# sourceMappingURL=index.d.ts.map